pyproject.toml
[project]
name = "spotify-llm"
version = "0.1.0"
description = "Agentic music discovery: chat + web search + Spotify."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Agent harness: https://github.com/langchain-ai/deepagents / https://docs.langchain.com/oss/python/deepagents/overview
    "deepagents>=0.4.0,<0.6",
    # Keep aligned with deepagents / LangChain supported range.
    "langgraph>=1.1.1,<1.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-tavily>=0.2.0",
    "spotipy>=2.23.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.6,<0.116",
    "starlette>=0.40.0,<0.42",
    "itsdangerous>=2.2.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.9",
    "langgraph-checkpoint-postgres>=2.0.0",
    "cryptography>=42.0.0",
]

[project.optional-dependencies]
playwright = ["playwright>=1.49.0", "langchain-community>=0.3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
